+++ angular/package.json
@@ -28,6 +28,7 @@
     "@angular/platform-browser-dynamic": "~12.2.4",
     "@angular/router": "~12.2.4",
     "@ngrx/effects": "^12.4.0",
+    "@ngrx/entity": "^12.4.0",
     "@ngrx/store": "^12.4.0",
     "bootstrap": "^4.5.3",
     "bootstrap-icons": "^1.3.0",
