--- angular/src/app/class/types.ts
@@ -1,11 +1,0 @@
-export interface User {
-    id: string;
-    username: string;
-    lastname: string;
-    firstname: string;
-    dateOfBirth: string;
-    email: string;
-    country: string;
-    region: string;
-    city: string;
-}
