+++ angular/src/app/login-page/services/authentification.service.ts
@@ -2,53 +2,63 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpResponse } from '@angular/common/http';
 import { Observable, throwError } from 'rxjs';
 import { catchError, map, retry, tap } from 'rxjs/operators';
-import { User } from './class/types';
+
 import { environment } from 'src/environments/environment';
+
+export interface User {
+  id: string;
+  username: string;
+  lastname: string;
+  firstname: string;
+  dateOfBirth: string;
+  email: string;
+  country: string;
+  region: string;
+  city: string;
+}
 
 const httpOptions = {
   headers: new HttpHeaders({
     'Content-Type': 'application/json',
   }),
-  withCredentials: true
+  withCredentials: true,
 };
 
-
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
-
 export class AuthentificationService {
   isLoggedIn = false;
   // store the URL so we can redirect after logging in
   redirectUrl: string;
   constructor(private http: HttpClient) {}
 
-
-  private authentificationUrl = environment.apiUrl + 'login-page/login';
-  private isAuthUrl = environment.apiUrl + 'login-page/is-auth';
-  private logoutUrl = environment.apiUrl + 'login-page/logout';
+  authentificationUrl = environment.apiUrl + 'login-page/login';
+  isAuthUrl = environment.apiUrl + 'login-page/is-auth';
+  logoutUrl = environment.apiUrl + 'login-page/logout';
   /**
    * login user
    */
   login(credentials: any): Observable<boolean | HttpResponse<User>> {
-    return this.http.post<User>(this.authentificationUrl, credentials, {
-      headers: new HttpHeaders({
-        'Content-Type': 'application/json',
-      }),
-      observe: 'response'
-    })
+    return this.http
+      .post<User>(this.authentificationUrl, credentials, {
+        headers: new HttpHeaders({
+          'Content-Type': 'application/json',
+        }),
+        observe: 'response',
+      })
       .pipe(
-        tap(res => {
+        tap((res) => {
           this.isLoggedIn = true;
         })
-      )
+      );
   }
   /**
    * logout
    */
   logout(): Observable<any> {
     return this.http.get<any>(this.logoutUrl, httpOptions).pipe(
-      map(res => {
+      map((res) => {
         this.isLoggedIn = res.authenticated;
         return this.isLoggedIn;
       })
@@ -59,11 +69,10 @@ export class AuthentificationService {
    */
   isAuth(): Observable<any> {
     return this.http.get<any>(this.isAuthUrl, httpOptions).pipe(
-      map(res => {
+      map((res) => {
         this.isLoggedIn = res.authenticated;
         return this.isLoggedIn;
       })
     );
   }
 }
-
