+++ angular/src/app/login-page/store/reducer.ts
@@ -1,0 +1,14 @@
+import { EntityState, EntityAdapter, createEntityAdapter } from '@ngrx/entity';
+
+export interface Authentification {
+  email: string;
+  password: string
+}
+
+export interface State extends EntityState<Authentification> {
+}
+
+
+export const adapter: EntityAdapter<Authentification> = createEntityAdapter<Authentification>({
+
+});
